refactor: #184 moved ExpansionValue type to common lib and moved Expansion.omitFromNamedCriteria() back to NamedCriteria.omitExpansion() (which was the main reason to generate the common lib in the first place; to resolve the cycli dependency between core + criteria
