+++ package.json
@@ -6,10 +6,10 @@
     "scripts": {
         "postinstall": "ngcc --properties es2015 browser module main",
         "test": "npm run test:types && npm run test:libs",
-        "test:libs": "npx nx run-many --parallel --target=test --projects=utils,lexer,criteria,core --silent --codeCoverage",
-        "test:types": "npm run test:types:criteria && npm run test:types:core",
-        "test:types:criteria": "dtslint --localTs node_modules/typescript/lib packages/criteria/src/test/types",
-        "test:types:core": "dtslint --localTs node_modules/typescript/lib packages/core/src/test/types"
+        "test:libs": "npx nx run-many --parallel --target=test --projects=utils,lexer,common,criteria,core --silent --codeCoverage",
+        "test:types": "npm run test:types:common && npm run test:types:criteria",
+        "test:types:common": "dtslint --localTs node_modules/typescript/lib packages/common/src/test/types",
+        "test:types:criteria": "dtslint --localTs node_modules/typescript/lib packages/criteria/src/test/types"
     },
     "repository": {
         "type": "git",
