+++ packages/common/src/index.ts
@@ -1,1 +1,2 @@
-export * from "./lib/common";
+export * from "./lib/entity.type";
+export * from "./lib/expansion-value.type";
