--- packages/common/src/lib/common.spec.ts
@@ -1,7 +1,0 @@
-import { common } from "./common";
-
-describe("common", () => {
-    it("should work", () => {
-        expect(common()).toEqual("common");
-    });
-});
