--- packages/common/src/lib/common.ts
@@ -1,3 +1,0 @@
-export function common(): string {
-    return "common";
-}
