+++ packages/common/src/lib/entity.type.ts
@@ -1,0 +1,1 @@
+export type Entity = Record<string, any>;
