+++ packages/common/src/test/types/tsconfig.json
@@ -9,10 +9,6 @@
         "strictNullChecks": true,
         "strictFunctionTypes": true,
         "noEmit": true,
-        "downlevelIteration": true,
-        "paths": {
-            "@entity-space/utils": ["../../../../utils/src/index.ts"],
-            "@entity-space/criteria": ["../../../../criteria/src/index.ts"]
-        }
+        "downlevelIteration": true
     }
 }
