+++ packages/core/src/lib/entity/functions/create-default-expansion.fn.ts
@@ -1,4 +1,4 @@
-import { ExpansionValue } from "../../expansion";
+import { ExpansionValue } from "@entity-space/common";
 import { IEntitySchema } from "../../schema";
 
 export function createDefaultExpansion(schema: IEntitySchema): ExpansionValue {
