+++ packages/core/src/lib/query/merge-queries.test.ts
@@ -1,5 +1,5 @@
+import { ExpansionValue } from "@entity-space/common";
 import { Criterion, inRange, matches, or } from "@entity-space/criteria";
-import { ExpansionValue } from "../expansion";
 import { EntitySchema } from "../schema";
 import { mergeQueries } from "./merge-queries.fn";
 import { Query } from "./query";
