+++ packages/core/src/lib/query/test/new-query-playground.test.ts
@@ -1,8 +1,9 @@
+import { ExpansionValue } from "@entity-space/common";
 import { Criterion, NamedCriteriaBag } from "@entity-space/criteria";
 import { Class } from "@entity-space/utils";
 import { Instance } from "../../entity/blueprint/instance";
 import { define } from "../../entity/blueprint/property";
-import { Expand, ExpansionValue } from "../../expansion";
+import { Expand } from "../../expansion";
 
 interface Brand {
     id: number;
