+++ packages/core/src/lib/query/test/reduce-queries.fn.test.ts
@@ -1,5 +1,5 @@
+import { ExpansionValue } from "@entity-space/common";
 import { Criterion, inSet, matches } from "@entity-space/criteria";
-import { ExpansionValue } from "../../expansion";
 import { EntitySchema } from "../../schema/entity-schema";
 import { Query } from "../query";
 import { reduceQueries } from "../reduce-queries.fn";
